+++ Xero.NetStandard.OAuth2/Model/Accounting/CISSetting.cs
@@ -43,7 +43,7 @@ namespace Xero.NetStandard.OAuth2.Model.Accounting
         /// </summary>
         /// <value>CIS Deduction rate for the contact if he is a subcontractor. If the contact is not CISEnabled, then the rate is not returned</value>
         [DataMember(Name="Rate", EmitDefaultValue=false)]
-        public decimal? Rate { get; private set; }
+        public double? Rate { get; private set; }
 
         /// <summary>
         /// Returns the string presentation of the object
